+++ alacritty_terminal/src/config/mod.rs
@@ -57,10 +57,6 @@ pub struct Config<T> {
     /// Bell configuration.
     bell: BellConfig,
 
-    #[cfg(windows)]
-    #[config(deprecated = "recompile with winpty feature or remove this setting")]
-    pub winpty_backend: bool,
-
     #[config(deprecated = "use `bell` instead")]
     pub visual_bell: Option<BellConfig>,
 }
@@ -135,7 +131,12 @@ impl Cursor {
 #[derive(Deserialize, Debug, Copy, Clone, PartialEq, Eq)]
 pub enum ConfigCursorStyle {
     Shape(CursorShape),
-    WithBlinking { shape: CursorShape, blinking: CursorBlinking },
+    WithBlinking {
+        #[serde(default)]
+        shape: CursorShape,
+        #[serde(default)]
+        blinking: CursorBlinking,
+    },
 }
 
 impl Default for ConfigCursorStyle {
@@ -199,7 +200,11 @@ impl Into<bool> for CursorBlinking {
 #[derive(Deserialize, Debug, Clone, PartialEq, Eq)]
 pub enum Program {
     Just(String),
-    WithArgs { program: String, args: Vec<String> },
+    WithArgs {
+        program: String,
+        #[serde(default)]
+        args: Vec<String>,
+    },
 }
 
 impl Program {